requirements.txt
Django==3.2.3
Cython
amqp
aniso8601
arrow
asgiref
async-timeout
atomicwrites
attrs
autobahn
Automat
Babel
backcall
billiard
celery
channels
channels-redis
chardet
constantly
coverage
daphne
decorator
django-cors-headers
django-crispy-forms
django-debug-toolbar===3.4.0
django-extensions
django-filter
django-mathfilters
django-phonenumber-field
django-split-settings
djangorestframework
drf_yasg
factory-boy
Faker
gunicorn
hyperlink
idna
incremental
ipython
ipython-genutils
jedi
kombu
Markdown
more-itertools
# numpy
# pandas
parso
pexpect
pickleshare
Pillow==9.0.1
pluggy
promise
prompt-toolkit
psycopg2
ptyprocess
py-moneyed
Pycco
Pygments
PyHamcrest
pystache
pytest
python-dateutil
pytz
sendgrid
singledispatch
six
smartypants
sqlparse
tblib
tex
traitlets
Twisted
txaio
uuid
vine
wcwidth
Werkzeug
zope.interface
